[MISC] 컴포넌트 내 함수 분리
